pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aya"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "discord.py[voice]>=2.3.2",
    "anthropic>=0.40.0",
    "voyageai>=0.3.2",
    "qdrant-client>=1.12.0",
    "redis>=5.0.0",
    "llama-index-core>=0.12.0",
    "structlog>=24.4.0",
    "pydantic-settings>=2.6.0",
    "numpy>=1.26.0",
    "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
